requirements.txt
Playwright
PyYAML
Pydantic
tqdm